Move code for loading from S3 into SaveLocalProperties
