+++ java/configuration/src/main/java/sleeper/configuration/properties/local/SaveLocalProperties.java
@@ -17,23 +17,27 @@
 package sleeper.configuration.properties.local;
 
 import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.S3ObjectSummary;
 
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.nio.file.Path;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.CONFIG_BUCKET;
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.QUERY_RESULTS_BUCKET;
 
 public class SaveLocalProperties {
     private final InstanceProperties instanceProperties;
-    private final LoadLocalProperties tables;
+    private final List<TableProperties> tables;
 
-    private SaveLocalProperties(InstanceProperties instanceProperties, LoadLocalProperties tables) {
+    private SaveLocalProperties(InstanceProperties instanceProperties, List<TableProperties> tables) {
         this.instanceProperties = instanceProperties;
         this.tables = tables;
     }
@@ -46,7 +50,7 @@ public class SaveLocalProperties {
             throw new UncheckedIOException(e);
         }
         return new SaveLocalProperties(instanceProperties,
-                LoadLocalProperties.loadFromS3(s3, instanceProperties));
+                loadTablesFromS3(s3, instanceProperties).collect(Collectors.toList()));
     }
 
     public static SaveLocalProperties loadFromPath(Path path) {
@@ -76,6 +80,27 @@ public class SaveLocalProperties {
     }
 
     public List<TableProperties> getTables() {
-        return tables.getTables();
+        return tables;
+    }
+
+    private static Stream<TableProperties> loadTablesFromS3(AmazonS3 s3, InstanceProperties instanceProperties) {
+        String configBucket = instanceProperties.get(CONFIG_BUCKET);
+        return s3.listObjectsV2(configBucket, "tables/")
+                .getObjectSummaries().stream()
+                .map(tableConfigObject -> loadTableFromS3(s3, instanceProperties, tableConfigObject));
+    }
+
+    private static TableProperties loadTableFromS3(
+            AmazonS3 s3, InstanceProperties instanceProperties, S3ObjectSummary tableConfigObject) {
+        TableProperties tableProperties = new TableProperties(instanceProperties);
+        try (InputStream in = s3.getObject(
+                        tableConfigObject.getBucketName(),
+                        tableConfigObject.getKey())
+                .getObjectContent()) {
+            tableProperties.load(in);
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+        return tableProperties;
     }
 }
